Fixed icon
Took 1 hour 30 minutes
