+++ app/src/main/AndroidManifest.xml
@@ -9,35 +9,24 @@
         android:roundIcon="@mipmap/logo_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme.NoActionBar">
-        <activity android:name=".Login.SplashActivity">
+        <activity android:name=".Ui.Login.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".Navigation.ui.Edit_Company_Info" />
+        <activity android:name=".Ui.Navigation.ui.Edit_Company_Info" />
         <activity android:name=".Individual_Product" />
         <activity
-            android:name=".Navigation.Navigation"
+            android:name=".Ui.Navigation.Navigation"
             android:label="@string/title_activity_navigation"
             android:theme="@style/AppTheme.NoActionBar.NoActionBar" />
-        <activity android:name=".InformationInput.Input" />
-        <activity android:name=".Login.verification" />
-        <activity android:name=".Login.Login">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
+        <activity android:name=".Ui.InformationInput.Input" />
+        <activity android:name=".Ui.Login.verification" />
+        <activity android:name=".Ui.Login.Login"/>
+        <activity android:name=".MainActivity"/>
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity android:name=".MainActivity">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
     </application>
 
 </manifest>
