+++ app/src/main/java/com/grace/profitabletraderconsultant/Ui/Navigation/ui/home/HomeViewModel.java
@@ -1,16 +1,16 @@
-package com.grace.profitabletraderconsultant.Navigation.ui.tools;
+package com.grace.profitabletraderconsultant.Ui.Navigation.ui.home;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class ToolsViewModel extends ViewModel {
+public class HomeViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public ToolsViewModel() {
+    public HomeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is tools fragment");
+        mText.setValue("This is home fragment");
     }
 
     public LiveData<String> getText() {
