+++ app/src/main/java/com/grace/profitabletraderconsultant/Ui/Navigation/ui/profile/ProfileViewModel.java
@@ -1,16 +1,16 @@
-package com.grace.profitabletraderconsultant.Navigation.ui.home;
+package com.grace.profitabletraderconsultant.Ui.Navigation.ui.profile;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class ProfileViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public ProfileViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is slideshow fragment");
     }
 
     public LiveData<String> getText() {
