+++ app/src/main/res/layout/activity_edit__company__info.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Navigation.ui.Edit_Company_Info">
+    tools:context=".Ui.Navigation.ui.Edit_Company_Info">
 
     <ImageButton
         android:id="@+id/back"
