+++ app/src/main/res/layout/activity_verification.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/White"
-    tools:context=".Login.verification">
+    tools:context=".Ui.Login.verification">
 
     <LinearLayout
         android:id="@+id/logo"
