+++ app/src/main/res/layout/app_bar_navigation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Navigation.Navigation">
+    tools:context=".Ui.Navigation.Navigation">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
