+++ app/src/main/res/layout/fragment_company_info.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".InformationInput.CompanyInfo">
+    tools:context=".Ui.InformationInput.CompanyInfo">
 
 
     <TextView
