+++ app/src/main/res/layout/fragment_product_info.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".InformationInput.ProductInfo">
+    tools:context=".Ui.InformationInput.ProductInfo">
 
     <TextView
         android:id="@+id/head"
